Add new method to fetch the number of seconds left in the day
This allows us to cache our data not for 24 hours, but for as many hours as there are left in the day.
This way, if the first visitor on the site on a given day visits the site at 11AM, we do not cache the posts until 11AM the next day. Instead we cache them until midnight on that day.